Switch up testing to rely on pytest which is the repo standard
